Init eslint
