+++ .eslintrc
@@ -1,0 +1,44 @@
+{
+  "parser": "babel-eslint",
+  "extends": "airbnb",
+  "env": {
+    // I write for browser
+    "browser": true,
+    // in CommonJS
+    "node": true,
+    "jest": true
+  },
+  "settings": {
+    "import/resolver": {
+      "node": {
+        "extensions": [".js", ".ios.js", ".android.js"]
+      }
+    }
+  },
+  "plugins": [
+    "react",
+    "react-native"
+  ],
+  "rules": {
+    "react/jsx-filename-extension": "off",
+    "react-native/no-unused-styles": "error",
+    "react-native/no-inline-styles": "error",
+    "react-native/no-color-literals": "error",
+    "react-native/split-platform-components": "error",
+    "consistent-this": ["error", "_this"],
+    "func-names": "error",
+    "max-len": ["error", 120, 2],
+    "no-console": "error",
+    "no-inline-comments": "error",
+    "no-lonely-if": "error",
+    "no-multiple-empty-lines": ["error", {
+      "max": 1,
+      "maxEOF": 1
+    }],
+    "no-param-reassign": ["error", {"props": false}],
+    "no-var": 2,
+    "operator-assignment": ["error", "always"],
+    "prefer-const": "error",
+    "quotes": ["error", "single"]
+  }
+}
