+++ src/components/AbstractComponent.js
@@ -1,0 +1,15 @@
+import { Component } from 'react';
+
+export default class AbstractComponent extends Component {
+  /**
+   * Constructor
+   * @param props
+   */
+  constructor(props) {
+    super(props);
+
+    if (this.constructor === AbstractComponent) {
+      throw new TypeError('Can not construct abstract class.');
+    }
+  }
+}
