+++ src/styles/ReactNativePlayground.theme.js
@@ -1,0 +1,15 @@
+/**
+ * Get theme colors constants
+ * @type {Object}
+ */
+export const COLORS = {
+  BACKGROUND_COLOR: '#F5FCFF',
+};
+
+/**
+ * Get theme constants
+ * @type {Object}
+ */
+export const THEME = {
+  COLORS,
+};
